Centralize initializeApp.
